Bckgrd color and border props in Wrapper
